Rename private modules
