+++ cropclassification/preprocess/_timeseries_calc_openeo.py
