+++ cropclassification/preprocess/_timeseries_calc_per_image.py
