+++ tests/test_calc_timeseries.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import sys
 
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
-from cropclassification.preprocess import timeseries_calc_dias_onda_per_image as calc_ts
+from cropclassification.preprocess import _timeseries_calc_per_image as calc_ts
 
 
 def get_testdata_dir() -> Path:
